--- .gitignore
@@ -1,27 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# dependencies
-src/frontend/node_modules
-src/backend/node_modules
-node_modules
-/.pnp
-.pnp.js
-
-
-# production
-/build
-
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-node_modules
-
-WebDesign-Project.code-workspace
