+++ README.md
@@ -1,10 +1,1 @@
-# WebDesign-Project
-
-Finally got the website to actually work and post at this point.
-How to start:
-Make sure you run npm i in both backend and front end folders.
-
-Then type the command 'npm run dev' in ./src/backend directory
-
-Then it should start and open on your page.
-
+SAMPLE
