--- package.json
@@ -1,68 +1,0 @@
-{
-  "name": "webdesign-project",
-  "version": "1.0.0",
-  "description": "Finally got the website to actually work and post at this point.\r How to start:\r Make sure you run npm i in both backend and front end folders.",
-  "main": "index.js",
-  "dependencies": {
-    "array-union": "^2.1.0",
-    "async": "^3.2.6",
-    "bcryptjs": "^2.4.3",
-    "braces": "^3.0.3",
-    "commander": "^11.1.0",
-    "commondir": "^1.0.1",
-    "cors": "^2.8.5",
-    "dir-glob": "^3.0.1",
-    "dotenv": "^16.4.5",
-    "email-addresses": "^5.0.0",
-    "escape-string-regexp": "^1.0.5",
-    "express": "^4.21.1",
-    "fast-glob": "^3.3.2",
-    "fastq": "^1.17.1",
-    "filename-reserved-regex": "^2.0.0",
-    "filenamify": "^4.3.0",
-    "fill-range": "^7.1.1",
-    "find-cache-dir": "^3.3.2",
-    "find-up": "^4.1.0",
-    "fs-extra": "^11.2.0",
-    "gh-pages": "^6.2.0",
-    "glob-parent": "^5.1.2",
-    "globby": "^11.1.0",
-    "graceful-fs": "^4.2.11",
-    "ignore": "^5.3.2",
-    "is-extglob": "^2.1.1",
-    "is-glob": "^4.0.3",
-    "is-number": "^7.0.0",
-    "jsonfile": "^6.1.0",
-    "jsonwebtoken": "^9.0.2",
-    "locate-path": "^5.0.0",
-    "make-dir": "^3.1.0",
-    "merge2": "^1.4.1",
-    "micromatch": "^4.0.8",
-    "mongodb": "^6.9.0",
-    "mongoose": "^8.7.2",
-    "p-limit": "^2.3.0",
-    "p-locate": "^4.1.0",
-    "p-try": "^2.2.0",
-    "path-exists": "^4.0.0",
-    "path-type": "^4.0.0",
-    "picomatch": "^2.3.1",
-    "pkg-dir": "^4.2.0",
-    "queue-microtask": "^1.2.3",
-    "reusify": "^1.0.4",
-    "run-parallel": "^1.2.0",
-    "semver": "^6.3.1",
-    "slash": "^3.0.0",
-    "strip-outer": "^1.0.1",
-    "to-regex-range": "^5.0.1",
-    "trim-repeated": "^1.0.0",
-    "universalify": "^2.0.1"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js",
-    "prestart": "npm run build --prefix webdesign"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
