+++ src/frontend/src/pages/App.css
@@ -1,0 +1,132 @@
+/* General page styling */
+body, html {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: 'Arial', sans-serif;
+  background-color: #f0f2f5;
+  height: 100%;
+}
+
+/* Reset all elements to use border-box for consistent sizing */
+*, *:before, *:after {
+  box-sizing: inherit;
+}
+
+/* Top bar styling */
+.top-bar {
+  background-color: #007bff;
+  padding: 15px 20px;
+  display: flex;
+  justify-content: flex-end; /* Align all items to the right */
+  align-items: center;
+  width: 100%;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 100;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+/* Navigation link styling (Home, Login, Signup, Dashboard) */
+.nav-link {
+  color: white;
+  text-decoration: none;
+  font-size: 1.2rem;
+  margin-left: 20px;
+  cursor: pointer;
+}
+
+.nav-link:hover {
+  text-decoration: underline;
+}
+
+/* Sign out button styling */
+.signout-button {
+  background: none;
+  border: none;
+  color: white;
+  font-size: 1.2rem;
+  margin-left: 20px;
+  cursor: pointer;
+  padding: 0;
+}
+
+.signout-button:hover {
+  text-decoration: underline;
+}
+
+/* Home page content styling */
+.home-content {
+  margin-top: 80px;
+  text-align: center;
+  padding: 20px;
+}
+
+.home-content h1 {
+  font-size: 2.5rem;
+  color: #333;
+}
+
+.home-content p {
+  font-size: 1.2rem;
+  color: #666;
+  max-width: 600px;
+  margin: 20px auto;
+  line-height: 1.6;
+}
+
+/* Button container styling */
+.button-container {
+  margin-top: 40px; /* Adds space between text and buttons */
+  display: flex;
+  justify-content: center;
+  gap: 20px; /* Space between buttons */
+}
+
+/* General button styling */
+.home-button {
+  padding: 15px 30px;
+  font-size: 1.5rem;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background-color 0.3s ease, transform 0.3s ease;
+}
+
+/* Signup button styling */
+.signup-button {
+  background-color: #28a745; /* Green button */
+  color: white;
+}
+
+.signup-button:hover {
+  background-color: #218838; /* Darker green on hover */
+  transform: scale(1.05); /* Slight zoom on hover */
+}
+
+/* Login button styling */
+.login-button {
+  background-color: #007bff; /* Blue button */
+  color: white;
+}
+
+.login-button:hover {
+  background-color: #0056b3; /* Darker blue on hover */
+  transform: scale(1.05); /* Slight zoom on hover */
+}
+
+/* Responsive styling */
+@media (max-width: 500px) {
+  .home-content h1 {
+    font-size: 2rem;
+  }
+
+  .home-content p {
+    font-size: 1rem;
+  }
+
+  .home-button {
+    font-size: 1.2rem;
+  }
+}
