--- webdesign/tsconfig.app.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/dashboard/dashboard.tsx","./src/dashboard/dashboard components/tasklist.tsx","./src/login/login.tsx","./src/signup/signup.tsx"],"version":"5.6.3"}
