--- webdesign/tsconfig.node.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./vite.config.ts"],"version":"5.6.3"}
